+++ packages/react/src/lib/components/LegendOverlay/Legend/index.ts
@@ -1,2 +1,3 @@
 export * from "./Legend";
 export * from "./LegendItem";
+export * from "./SizeLegend";